backend-repo/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
openai>=1.40.0
python-dotenv
